refactor(deck_research_panel): split DeckResearchPanel into mixin package
Move widgets/panels/deck_research_panel.py into a package: frame.py keeps
the _CenteredChoice helper and the full _build_ui; properties.py holds
the pure filter getters; handlers.py owns the UI-touching setters,
resets, loading/error states, archetype population, and enable/disable
toggles.
