+++ widgets/panels/deck_research_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Deck research panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.deck_research_panel.frame import DeckResearchPanel
+
+__all__ = ["DeckResearchPanel"]
